backend/xmrt-dao-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitarray==3.4.3
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
ckzg==2.1.1
click>=8.0.0,<9.0.0
cytoolz==1.0.1
distro==1.9.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.6.3
openai==1.93.0
parsimonious==0.10.0
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
web3==7.12.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
